chore: Adds proxy to jupyterhub via nginx and fileglancer-central hostname
This sets up the proxy to jupyterhub, which is needed to serve it via an
https connection. In order to have both the jupyterhub and
fileglancer-central services listening on port 443, they have been split
via namebased routing on the nginx server. The jupyterhub content will
be served via the fileglancer-dev domain and the fileglancer-central
service will be served via a domain of the same name.