fix: vastly simplifies migration script.
The use of the alembic framework to build the schema in the postgres
target was causing schema mismatches and failed alembic migrations on
server start up. This simplified version copies the existing data, so
that it should be in the exact same state after migration. This way
alembic should see the exact same schema and alembic_Version in the
database and not fail to migrate the data.